+++ src/main.js
@@ -1,6 +1,7 @@
 import { NoteModel } from './models/note_model';
 import { NoteSheet } from './ui/note_sheet';
 import { initializeNotesTab } from './ui/notes_tab';
+import { registerSearchEnhancement } from './ui/search_enhancement';
 
 Hooks.once('init', () => {
   console.log('5e Sheet Notes & Trackers | Module initialized');
@@ -17,4 +18,8 @@ Hooks.once('init', () => {
   });
 
   initializeNotesTab();
+});
+
+Hooks.once('setup', () => {
+  registerSearchEnhancement();
 });